benchmarks/requirements.txt
paddlepaddle==2.4.0rc0
paddle-quantum==2.2.0
